+++ CHANGELOG.md
@@ -4,6 +4,15 @@
   
     *Megan O'Neill*
 
+## 0.8.3 (May 7, 2018)
+
+## 0.8.2 (May 7, 2018)
+
+* Rescue Procore::OAuthError
+* Add Procore::MissingTokenError
+
+    *Casey Ochs*
+
 ## 0.8.1 (April 13, 2018)
 
 * Fix rubocop